rename patched llamaindex stuff to be clearer
