+++ src/main/services/LoggingVectorStoreIndex.ts
