+++ src/main/services/batchingWeaviateVectorStore.ts
@@ -1,7 +1,15 @@
 import { WeaviateVectorStore } from '@llamaindex/weaviate';
 import { BaseNode } from 'llamaindex';
 
-export class PatchedWeaviateVectorStore extends WeaviateVectorStore {
+/*
+
+Patched version of WeaviateVectorStore to handle large batches by splitting into smaller chunks.
+
+When I loaded a large-ish (5.4MB) spreadsheet, I got a Weaviate error about trying to load too much data at once.
+
+*/
+
+export class BatchingWeaviateVectorStore extends WeaviateVectorStore {
   async add(nodes: BaseNode[]): Promise<string[]> {
     const batchSize = 100; // Define the batch size
     const results: string[] = []; // Collect results from each batch
